# perf: eliminate omit and reduce intersections to improve file based routing typescript performance