A redundant intersection `UpdatableRouteOption`  in file routes was causing a lot more instantiations when there are many file based routes. Intersections recursively merge properties so if there are many of them then this can slow down typescript, Turns out route options in general create a lot of intersections.

I've reduced unnecessary intersections.

Use separate type alias instead of Omit. This further reduces instantiations.

I have 256 file routes.

Extended diagnostics before:

```
Files:                         444
Lines of Library:            38118
Lines of Definitions:        77729
Lines of TypeScript:         10338
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                114975
Symbols:                    260848
Types:                       75313
Instantiations:             801208
Memory used:               269042K
Assignability cache size:    32031
Identity cache size:          3460
Subtype cache size:              2
Strict subtype cache size:   19640
I/O Read time:               0.07s
Parse time:                  0.57s
ResolveModule time:          0.11s
ResolveTypeReference time:   0.01s
ResolveLibrary time:         0.02s
Program time:                0.86s
Bind time:                   0.22s
Check time:                  3.40s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                  4.48s
```

Extended diagnostics after:

Instantiations are almost halved

```
Files:                         444
Lines of Library:            38118
Lines of Definitions:        77729
Lines of TypeScript:         10338
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                115006
Symbols:                    179907
Types:                       54277
Instantiations:             403090
Memory used:               213675K
Assignability cache size:    32842
Identity cache size:          3460
Subtype cache size:              2
Strict subtype cache size:   19640
I/O Read time:               0.06s
Parse time:                  0.52s
ResolveModule time:          0.10s
ResolveTypeReference time:   0.00s
ResolveLibrary time:         0.02s
Program time:                0.78s
Bind time:                   0.22s
Check time:                  2.64s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                  3.65s
```